+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - `ItemDefindex` trait for definitions associated with an item defindex.
 - `IntoEnumIterator`, `EnumCount`, `TryFromPrimitive`, `IntoPrimitive` re-exports.
 - Missing `From` and `TryFrom` conversions for spell enums.
+- `is_paint_spell` and `is_footprints_spell` methods to `Spell`.
 
 ### Removed
 - `defindex_is_strange_part` for `StrangePart`.